refactor(TapDetector): Generalisation and Abstraction of TapDetector Interface and Logic
This allows to implement detectors that detect tap differently.
